Fixed RMSE for Equation D
Fixed the RMSE for equation D. All RMSE's are now reading correctly.
Unusually high error in RMSE of equation B. The result is a high RMSE
for equation D. Requires immediate debugging. NOTE: if you can't see the
output it is because i'm a noob at printing in python. I tried flushing
the system's stdout with no luck. Just break and F9 through it.